Move domain classes to one file
