--- game/level/platformKind.h
@@ -1,11 +1,0 @@
-//
-// Created by Joris on 26-9-2018.
-//
-
-#pragma once
-namespace levelDomain {
-    enum class PlatformKind {
-        Solid = 0,
-        UnderPassable = 1
-    };
-}
