--- game/level/themes.h
@@ -1,9 +1,0 @@
-#pragma once
-
-enum class Theme {
-    Earth = 0,
-    Moon = 1,
-    Temple = 2,
-    City = 3
-
-};
