--- game/level/tile.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include <string>
-#include "external/JSON/json.hpp"
-using json = nlohmann::json;
-#include "platformKind.h"
-
-namespace levelDomain {
-    struct tile {
-        int x;
-        int y;
-        std::string sprite;
-        PlatformKind kind;
-    };
-}
-
